Complete day 09 Updated intcode computer to support relative parameter mode, long integers and memory that exceeds the initial program size

+++ src/de/ungefroren/adventofcode/y2019/intcodeComputer/instructions/JumpIfFalse.scala
@@ -9,7 +9,7 @@ object JumpIfFalse extends Instruction(6,2) {
 
   override def execute(param: Seq[Int], computer: IntcodeComputer): ComputationResult = {
     if (computer.memory(param.head) == 0) {
-      ComputationResult.JUMPTO(computer.memory(param(1)))
+      ComputationResult.JUMPTO(computer.memory(param(1)).toInt)
     } else {
       ComputationResult.CONTINUE_EXECUTION
     }
